backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# RAG & ML Dependencies
chromadb==0.4.18
sentence-transformers==2.3.1
huggingface-hub>=0.20.0
groq==0.4.1
PyPDF2==3.0.1

# Frontend
streamlit==1.28.1
requests==2.31.0

# Additional utilities
pydantic==2.5.0
pydantic-settings==2.1.0